feat: build both sdk versions
