+++ packages/crypto/package.json
@@ -13,9 +13,9 @@
     "release:dry-run": "yarn prepublish && release-it --verbose --dry-run --ci",
     "release:no-npm": "yarn prepublish && release-it --verbose --no-npm.publish --ci",
     "wasm:build": "./scripts/build.sh --release",
-    "wasm:build:dev": "./scripts/build-node.sh",
-    "wasm:build:node": "./scripts/build-node.sh",
-    "wasm:build:node:dev": "./scripts/build.sh",
+    "wasm:build:dev": "./scripts/build.sh",
+    "wasm:build:node": "./scripts/build-node.sh --release",
+    "wasm:build:node:dev": "./scripts/build-node.sh",
     "test": "yarn wasm:build:node && yarn jest",
     "test:watch": "yarn wasm:build:node && yarn test --watch",
     "test:ci": "jest",
