+++ packages/sdk/package.json
@@ -2,14 +2,21 @@
   "name": "@namada/sdk",
   "version": "0.2.4",
   "description": "Namada SDK package",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "exports": {
+    "./nodejs": "./dist/indexNode.js",
+    "./web": "./dist/indexWeb.js"
+  },
   "repository": "https://github.com/anoma/namada-interface/",
   "author": "Heliax Dev <info@heliax.dev>",
   "license": "MIT",
   "scripts": {
-    "bundle-dts": "dts-bundle-generator --no-check --project tsconfig.webpack.json -o dist/index.d.ts dist/dts/sdk/src/index.d.ts && rm -rf dist/dts",
-    "build": "rm -rf ./dist && yarn wasm:build:node:dev && webpack-cli && yarn bundle-dts",
+    "prebuild": "rimraf ./dist",
+    "build:web": "yarn wasm:build:dev && webpack-cli --config webpack.web.js",
+    "build:node": "yarn wasm:build:node:dev && webpack-cli --config webpack.node.js",
+    "build": "yarn prebuild && yarn build:web && yarn build:node && yarn bundle-dts",
+    "bundle-dts:web": "dts-bundle-generator --no-check --project tsconfig.webpack.json -o dist/indexWeb.d.ts dist/dts/sdk/src/indexWeb.d.ts",
+    "bundle-dts:node": "dts-bundle-generator --no-check --project tsconfig.webpack.json -o dist/indexNode.d.ts dist/dts/sdk/src/indexNode.d.ts",
+    "bundle-dts": "yarn bundle-dts:web && yarn bundle-dts:node && rimraf dist/dts",
     "example": "ts-node -r tsconfig-paths/register ./examples/index.ts",
     "lint": "eslint src --ext .ts,.tsx",
     "lint:fix": "yarn lint -- --fix",
