+++ packages/sdk/src/index.ts
@@ -1,7 +1,3 @@
-import { webcrypto } from "node:crypto";
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-(globalThis as any).crypto = webcrypto;
-
 // Make Ledger available for direct-import as it is not dependent on Sdk initialization
 export * from "./ledger";
 
@@ -17,6 +13,3 @@ export type {
   Unbonds,
 } from "./rpc";
 export { EncodedTx, SignedTx } from "./tx";
-
-// Export init functions for direct usage
-export { default as initSync } from "./initSync";
