+++ packages/sdk/src/indexWeb.ts
@@ -1,0 +1,3 @@
+export * from "./index";
+
+export { default as initAsync } from "./initAsync";
